Unshim GpuRowBasedScalaUDF
